+++ docs/README.md
@@ -6,7 +6,7 @@
 
 - [x] (e) 로또 번호의 개수가 6 이상이면 예외가 발생한다.
 - [x] (e) 로또 번호에 중복된 숫자가 있으면 예외가 발생한다.
-- [ ] 당첨번호를 오름차순으로 정렬한다.
+- [x] 당첨번호를 오름차순으로 정렬한다.
 
 ### Lottos
 
@@ -31,8 +31,8 @@
 ### TicketIssuer
 
 - [ ] TicketBooth에게 구매 수량을 전달 받는다.
-- [x] 수량에 해당하는 만큼 로또를 발행한다.
-  - [x] 1~45 범위의 중복되지 않는 숫자 6개를 생성한다.
+- [ ] 수량에 해당하는 만큼 로또를 발행한다.
+  - [ ] 1~45 범위의 중복되지 않는 숫자 6개를 생성한다.
   - [ ] 발행된 로또를 Lotto에 담는다.
   - [ ] Lottos 객체로 포장한다.
